Update swap example
