+++ examples/ts-sdk/next/next.config.js
@@ -1,50 +1,26 @@
 import NextBundleAnalyzer from "@next/bundle-analyzer";
 import CopyWebpackPlugin from "copy-webpack-plugin";
-import fs from "fs";
-import path from "path";
 
 const nextConfig = {
   serverExternalPackages: ["@orca-so/whirlpools-core"],
   webpack(config, { isServer }) {
     config.experiments.asyncWebAssembly = true;
 
-    // Locate WASM file for Orca Whirlpools Core
-    const nodeModulesWasmPath =
-      "node_modules/@orca-so/whirlpools-core/dist/nodejs/orca_whirlpools_core_js_bindings_bg.wasm";
-    const relativeWasmPath =
-      "../../../ts-sdk/core/dist/nodejs/orca_whirlpools_core_js_bindings_bg.wasm";
-
-    let wasmPath;
-    if (fs.existsSync(path.resolve(nodeModulesWasmPath))) {
-      wasmPath = nodeModulesWasmPath;
-    } else if (fs.existsSync(path.resolve(relativeWasmPath))) {
-      wasmPath = relativeWasmPath;
-    } else {
-      throw new Error(
-        "Could not find orca_whirlpools_core_js_bindings_bg.wasm file",
-      );
-    }
-
+    // Copy `orca_whirlpools_core_js_bindings_bg.wasm` file
+    // This is only needed because of the monorepo setup
+    // (local dependencies are symlinked and next doesn't like that)
     config.plugins.push(
       new CopyWebpackPlugin({
         patterns: [
           {
-            from: wasmPath,
-            to: "./server/app/swap",
-          },
-          {
-            from: wasmPath,
+            from: "../../../ts-sdk/core/dist/nodejs/orca_whirlpools_core_js_bindings_bg.wasm",
             to: "./server/app",
-          },
-          {
-            from: wasmPath,
-            to: "./server/chunks",
           },
         ],
       }),
     );
 
-    // Enable async functions in client-side bundles
+    // The following supresses a warning about using top-level-await and is optional
     if (!isServer) {
       config.output.environment = {
         ...config.output.environment,
